requirements.txt
# Updated requirements.txt with compatible LangChain versions
# Core dependencies with version compatibility fixes

# LangChain Core (pin to compatible version)
langchain-core==0.1.52
langchain==0.1.20
langchain-community==0.0.38

# LangChain Google GenAI (compatible version)
langchain-google-genai==1.0.3

# Google AI dependencies
google-generativeai==0.4.1
google-cloud-aiplatform==1.38.1

# Vector store and embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# Document processing
python-docx==1.1.0
python-dotenv==1.0.0

# Streamlit and UI
streamlit==1.31.1
plotly==5.18.0
pandas==2.1.4

# Utilities
pathlib
datetime
asyncio
typing-extensions==4.9.0

# Optional: Alternative packages if issues persist
# langchain-openai==0.0.8  # Alternative LLM
# faiss-cpu==1.7.4  # Alternative vector stores